chore: auto jump
